command+Vで間違えてしまった分
